requirements.txt
requests~=2.32.3
pytest==8.3.3
allure-pytest~=2.13.5
allure-python-commons~=2.13.5
jsonschema~=4.23.0